complete moving utility methods to same file
