+++ bigraph_schema/registry.py
@@ -17,31 +17,6 @@ from pprint import pformat as pf
 from bigraph_schema.protocols import local_lookup_module, function_module
 
 
-NONE_SYMBOL = '!nil'
-
-
-required_schema_keys = set([
-    '_default',
-    '_apply',
-    '_check',
-    '_serialize',
-    '_deserialize',
-    '_fold',
-])
-
-
-optional_schema_keys = set([
-    '_type',
-    '_value',
-    '_description',
-    '_type_parameters',
-    '_inherit',
-    '_divide',
-])
-
-
-type_schema_keys = required_schema_keys | optional_schema_keys
-
 
 def deep_merge_copy(dct, merge_dct):
     return deep_merge(copy.deepcopy(dct), merge_dct)
