+++ bigraph_schema/type_system.py
@@ -13,12 +13,12 @@ from pprint import pformat as pf
 
 from bigraph_schema import Registry, non_schema_keys, is_schema_key, deep_merge, type_parameter_key
 from bigraph_schema.parse import parse_expression
-from bigraph_schema.registry import type_schema_keys, remove_omitted, set_path, transform_path
+from bigraph_schema.utilities import union_keys, type_schema_keys
+from bigraph_schema.registry import remove_omitted, set_path, transform_path
 from bigraph_schema.type_functions import (
     registry_types, base_types, unit_types,
     register_base_reactions, is_empty,
     apply_schema, set_apply)
-from bigraph_schema.utilities import union_keys
 
 TYPE_FUNCTION_KEYS = [
     '_apply',
